Remove google-cloud specific unit tests in data_design_test, data_set_test, experimental_trial_test, and experimental_design_test upon Matthew's suggestion.
