+++ src/driver/tests/experimental_design_test.py
@@ -13,20 +13,13 @@
 # limitations under the License.
 """Tests for ExperimentalDesign."""
 
+from typing import Dict, Iterable, List, Type
 from absl.testing import absltest
 from os.path import join
 from tempfile import TemporaryDirectory
-from typing import Dict
-from typing import Iterable
-from typing import List
-from typing import Type
 import numpy as np
 import pandas as pd
-from unittest.mock import patch
 import logging
-
-import pathlib
-import cloudpathlib.local
 
 import apache_beam as beam
 from apache_beam.options.pipeline_options import PipelineOptions
@@ -73,7 +66,6 @@ from wfa_planning_evaluation_framework.driver.experiment_parameters import (
 from wfa_planning_evaluation_framework.driver.experiment import (
     Experiment,
 )
-import wfa_planning_evaluation_framework.driver.experimental_design as experimental_design
 from wfa_planning_evaluation_framework.driver.experimental_design import (
     ExperimentalDesign,
     CombineDataFrameFn,
@@ -91,12 +83,6 @@ from wfa_planning_evaluation_framework.driver.test_point_generator import (
 )
 from wfa_planning_evaluation_framework.driver.trial_descriptor import (
     TrialDescriptor,
-)
-from wfa_planning_evaluation_framework.filesystem_wrappers import (
-    filesystem_pathlib_wrapper,
-)
-from wfa_planning_evaluation_framework.filesystem_wrappers import (
-    filesystem_cloudpath_wrapper,
 )
 
 
@@ -153,41 +139,7 @@ class FakeExperimentalTrial:
         return self._base * seed
 
 
-class FakeEvaluateTrialDoFn(beam.DoFn):
-    def process(self, trial, seed, filesystem):
-        import pandas as pd
-
-        yield pd.DataFrame({"col": [1]})
-
-
-def fake_open(
-    self,
-    path,
-    mode="r",
-    buffering=-1,
-    encoding=None,
-    errors=None,
-    newline=None,
-):
-    if path.startswith("gs://"):
-        path = cloudpathlib.local.LocalGSPath(path)
-    else:
-        path = pathlib.Path(path)
-
-    return path.open(
-        mode=mode,
-        buffering=buffering,
-        encoding=encoding,
-        errors=errors,
-        newline=newline,
-    )
-
-
 class ExperimentalDesignTest(absltest.TestCase):
-    def tearDown(self):
-        cloudpathlib.local.localclient.clean_temp_dirs()
-        cloudpathlib.local.LocalGSClient.reset_default_storage_dir()
-
     def _setup(self, tempdir):
         pdf1 = PublisherData([(1, 0.01), (2, 0.02), (1, 0.04), (3, 0.05)], "pdf1")
         data_set1 = DataSet([pdf1], "dataset1")
@@ -268,60 +220,14 @@ class ExperimentalDesignTest(absltest.TestCase):
             # been checked by the preceding unit tests.  However, we make a few strategic
             # probes.
             self.assertEqual(results.shape[0], 1)
-            self.assertAlmostEqual(results["relative_error_at_100"][0], 0.0, delta=0.01)
+            self.assertAlmostEqual(results["relative_error_at_100"][0], 0.0, delta=0.02)
             self.assertGreater(results["max_nonzero_frequency_from_halo"][0], 0)
             self.assertEqual(results["max_nonzero_frequency_from_data"][0], 1)
-
-    @patch.object(experimental_design, "EvaluateTrialDoFn", FakeEvaluateTrialDoFn)
-    @patch.object(
-        filesystem_cloudpath_wrapper,
-        "CloudPath",
-        cloudpathlib.local.LocalGSPath,
-    )
-    @patch.object(
-        filesystem_cloudpath_wrapper.FilesystemCloudpathWrapper,
-        "set_default_client_to_gs_client",
-        cloudpathlib.local.LocalGSClient.get_default_client,
-    )
-    @patch.object(
-        filesystem_cloudpath_wrapper.FilesystemCloudpathWrapper,
-        "open",
-        fake_open,
-    )
-    def test_evaluate_with_apache_beam_with_cloud_path(self):
-        # Client setup
-        client = cloudpathlib.local.LocalGSClient.get_default_client()
-
-        filesystem = filesystem_cloudpath_wrapper.FilesystemCloudpathWrapper()
-
-        parent_dir_path = client.CloudPath("gs://ExperimentalDesignTest/parent")
-        data_design_dir_path = parent_dir_path.joinpath("data_design")
-        experiment_dir_path = parent_dir_path.joinpath("experiments")
-        data_design_dir_path.joinpath("dummy.txt").write_text(
-            "For creating the target directory."
-        )
-        experiment_dir_path.joinpath("dummy.txt").write_text(
-            "For creating the target directory."
-        )
-
-        results = self._evaluate_with_apache_beam(
-            data_design_dir=str(data_design_dir_path),
-            experiment_dir=str(experiment_dir_path),
-            use_cloud_path=True,
-            filesystem=filesystem,
-        )
-        # We don't check each column in the resulting dataframe, because these have
-        # been checked by the preceding unit tests.  However, we make a few strategic
-        # probes.
-        self.assertEqual(results.shape[0], 1)
-        self.assertEqual(results["col"][0], 1)
 
     def _evaluate_with_apache_beam(
         self,
         data_design_dir,
         experiment_dir,
-        use_cloud_path=False,
-        filesystem=filesystem_pathlib_wrapper.FilesystemPathlibWrapper(),
     ):
         num_workers = 2
         data1 = HeterogeneousImpressionGenerator(
@@ -329,7 +235,7 @@ class ExperimentalDesignTest(absltest.TestCase):
         )()
         pdf1 = PublisherData(FixedPriceGenerator(0.1)(data1))
         data_set = DataSet([pdf1], "dataset")
-        data_design = DataDesign(data_design_dir, filesystem)
+        data_design = DataDesign(data_design_dir)
         data_design.add(data_set)
 
         msd = ModelingStrategyDescriptor(
@@ -352,24 +258,15 @@ class ExperimentalDesignTest(absltest.TestCase):
             seed=1,
             cores=num_workers,
             analysis_type="single_pub",
-            filesystem=filesystem,
         )
         exp.generate_trials()
-
-        temp_location = None
-        if use_cloud_path:
-            client = cloudpathlib.local.LocalGSClient.get_default_client()
-            experiment_dir_cloud_path = client.CloudPath(experiment_dir)
-            temp_location = client._cloud_path_to_local(experiment_dir_cloud_path)
-        else:
-            temp_location = experiment_dir
 
         pipeline_args = []
         pipeline_args.extend(
             [
                 "--runner=direct",
                 "--direct_running_mode=multi_processing",
-                f"--temp_location={temp_location}",
+                f"--temp_location={experiment_dir}",
                 f"--direct_num_workers={num_workers}",
             ]
         )
